convert parallel test to sequential
