+++ core/src/test/resources/templates/analysis-mode-test/settings.gradle
@@ -24,7 +24,5 @@
 
 rootProject.name = 'analysis-mode-test'
 include 'Target'
-include 'DepA'
-include 'DepB'
-include 'DepC'
+include 'Dep'
